final project/code/requirements.txt
selenium
pytest
pytest-xdist
allure-pytest
webdriver-manager
pymysql
faker
requests
urllib
sqlalchemy
seleniumwrapper